requirements.txt
dnspython
pyfiglet
termcolor
requests
folium
ipwhois
matplotlib
